# problem-5/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Animal CRUD API",
    "description": "A complete CRUD API for managing animal resources",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "tags": [
    {
      "name": "Animals",
      "description": "Animal management endpoints"
    }
  ],
  "components": {
    "schemas": {
      "Animal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier"
          },
          "name": {
            "type": "string",
            "description": "Animal name"
          },
          "species": {
            "type": "string",
            "description": "Animal species"
          },
          "breed": {
            "type": "string",
            "description": "Animal breed"
          },
          "age": {
            "type": "integer",
            "description": "Animal age in years",
            "minimum": 0,
            "maximum": 50
          },
          "size": {
            "type": "string",
            "enum": [
              "small",
              "medium",
              "large"
            ],
            "description": "Animal size category"
          },
          "location": {
            "type": "string",
            "description": "Animal location"
          },
          "description": {
            "type": "string",
            "description": "Additional description"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "name",
          "species",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateAnimalRequest": {
        "type": "object",
        "required": [
          "name",
          "species"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Animal name"
          },
          "species": {
            "type": "string",
            "description": "Animal species"
          },
          "breed": {
            "type": "string",
            "description": "Animal breed"
          },
          "age": {
            "type": "integer",
            "description": "Animal age in years",
            "minimum": 0,
            "maximum": 50
          },
          "size": {
            "type": "string",
            "enum": [
              "small",
              "medium",
              "large"
            ],
            "description": "Animal size category"
          },
          "location": {
            "type": "string",
            "description": "Animal location"
          },
          "description": {
            "type": "string",
            "description": "Additional description"
          }
        }
      },
      "UpdateAnimalRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Animal name"
          },
          "species": {
            "type": "string",
            "description": "Animal species"
          },
          "breed": {
            "type": "string",
            "description": "Animal breed"
          },
          "age": {
            "type": "integer",
            "description": "Animal age in years",
            "minimum": 0,
            "maximum": 50
          },
          "size": {
            "type": "string",
            "enum": [
              "small",
              "medium",
              "large"
            ],
            "description": "Animal size category"
          },
          "location": {
            "type": "string",
            "description": "Animal location"
          },
          "description": {
            "type": "string",
            "description": "Additional description"
          }
        }
      },
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Operation success status"
          },
          "data": {
            "description": "Response data"
          },
          "message": {
            "type": "string",
            "description": "Success message"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Response timestamp"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Operation success status"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Animal"
            },
            "description": "Array of animals"
          },
          "pagination": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "description": "Total number of animals"
              },
              "limit": {
                "type": "integer",
                "description": "Number of results per page"
              },
              "offset": {
                "type": "integer",
                "description": "Number of results skipped"
              },
              "hasMore": {
                "type": "boolean",
                "description": "Whether there are more results"
              }
            },
            "required": [
              "total",
              "limit",
              "offset",
              "hasMore"
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Response timestamp"
          }
        },
        "required": [
          "success",
          "data",
          "pagination",
          "timestamp"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Operation success status"
          },
          "error": {
            "type": "string",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Response timestamp"
          },
          "path": {
            "type": "string",
            "description": "Request path"
          }
        },
        "required": [
          "success",
          "error",
          "message",
          "timestamp"
        ]
      }
    }
  },
  "paths": {
    "/api/animals": {
      "post": {
        "tags": [
          "Animals"
        ],
        "summary": "Create a new animal",
        "description": "Creates a new animal with the provided data",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnimalRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Animal created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Animals"
        ],
        "summary": "List animals with filters",
        "description": "Retrieves a list of animals with optional filtering",
        "parameters": [
          {
            "in": "query",
            "name": "species",
            "schema": {
              "type": "string"
            },
            "description": "Filter by species"
          },
          {
            "in": "query",
            "name": "breed",
            "schema": {
              "type": "string"
            },
            "description": "Filter by breed"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Number of results"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Number of results to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "List of animals",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/animals/{id}": {
      "get": {
        "tags": [
          "Animals"
        ],
        "summary": "Get animal by ID",
        "description": "Retrieves a specific animal by its ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Animal ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Animal details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Animals"
        ],
        "summary": "Update animal",
        "description": "Updates an existing animal",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Animal ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAnimalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Animal updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Animals"
        ],
        "summary": "Delete animal",
        "description": "Deletes an animal by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Animal ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Animal deleted"
          }
        }
      }
    }
  }
}